Backend/requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
python-jose
pymongo
python-dotenv
